storyboard_engine/requirements.txt
replicate
google-generativeai
watchdog
python-dotenv
requests
Pillow
numpy
